+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KX57CTSDEZTJE4V96QWPPCVS"
+        "01KX5DJEHN432FF5N2HV0RM52T"
     )
     r"""The unique identifier of the data source"""