+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KX57CTSHPRC5Q9P3YWVN2F8T"
+        "01KX5DJEHSGQRJZJFTJNRC51QM"
     )
     r"""The unique identifier of the data source"""