+++ packages/orq-rc/RELEASES.md
@@ -9399,3 +9399,13 @@ Based on:
 - [python v4.12.0-rc.12] packages/orq-rc
 ### Releases
 - [PyPI v4.12.0-rc.12] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.12 - packages/orq-rc
+
+## 2026-07-10 07:08:20
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.789.1 (2.916.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.12.0-rc.13] packages/orq-rc
+### Releases
+- [PyPI v4.12.0-rc.13] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.13 - packages/orq-rc